silence warning

+++ test_fast_map.h
@@ -130,7 +130,7 @@ TEST_F(AFastMap, CanBeRebuiltFully)
 
 	m_map.full_rebuild();
 
-	EXPECT_EQ(count, m_map.size());
+	EXPECT_EQ((size_t)count, m_map.size());
 	
 	for (const auto& pair : pairs)
 	{
